+++ lib/src/main/java/com/wjduquette/joe/walker/WalkerEngine.java
@@ -1,6 +1,7 @@
 package com.wjduquette.joe.walker;
 
 import com.wjduquette.joe.*;
+import com.wjduquette.joe.parser.Dumper;
 import com.wjduquette.joe.parser.Parser;
 import com.wjduquette.joe.parser.Stmt;
 import com.wjduquette.joe.SourceBuffer;
